CLI.py: fix block device udev race condition
We are encouraged to add a udev rule to change a block device's
bdi/max_ratio to '1' and queue/scheduler to 'none', which I did.
So I was surprised when, about 50% of the time, I kept seeing:

	...
	bmaptool: info: failed to enable I/O optimization, expect suboptimal speed (reason: cannot switch to the 'none' I/O scheduler: 'bfq' in use. [Errno 13] Permission denied: '/sys/dev/block/8:160/queue/scheduler')
	bmaptool: info: You may want to set these I/O optimizations through a udev rule like this:
	...

The strange part is that sometimes it doesn't report a problem and
sometimes it does, even if the block device is left plugged in continuously
between multiple bmaptool invocations.

In all of my tests the bdi/max_ratio is always okay, but the
queue/scheduler would sometimes be reported as being the default scheduler,
not the one the udev rule was setting (none). Yet no matter how many times
I would read the file outside of bmaptool it always would be set to the
correct scheduler.

It turns out that opening a block device in "wb+" mode, which is what
bmaptool is doing at one point, causes the block device to act as though
it was just inserted, giving it the default settings, then causing udev to
trigger to switch it to the requested settings. However, if udev doesn't
finish before bmaptool reads the scheduler value there's a chance it will
read the pre-udev value, not the post-udev value, even though the block
device was never physically removed and re-inserted.

bmaptool was opening every file, then checking for block devices and
if found, closing then re-opening the block devices via a special
block-opening helper function. This patch re-organizes the code to only
open block devices once using the special block-opening helper function
that does not open block devices in "wb+" mode.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>