add --keyring and --fingerprint options
These options allow one to pass a custom keyring with --keyring as well
as panning the fingerprint using --fingerprint. This is for example
useful in situations where a vendor wants to give users access to
updated system images and for that purpose, ships a custom keyring with
their installations. Using the --keyring option, the user can use GPG to
verify that the updated system image they downloaded indeed comes from
the expected source (even more so when using the --fingerprint option
together with it) without having to add the key into their own personal
GPG keyring.

Furthermore, this is useful in scripts which call bmaptool
programmatically. For example, an updater script could come shipped with
an embedded GPG keyring and then call bmaptool with this keyring (and
optionally with a known --fingerprint) to facilitate GPG-verified
automatic updates.