Resolve symlinks in device path for removable check
When running with --removable-device, the basename of dest is used to
construct the path of the "removable" sysfs file for the
device. However, if dest is a symlink the basename may be very
different from the actual device name, and theoretically might even
match a different device. Make the path absolute and resolve any
symlinks before taking the basename to avoid that problem.

This is of practical importance because it allows using
/dev/disk/... symlinks created by udev, which is much less error-prone
than using e.g. /dev/sdX.

Signed-off-by: Fiona Klute <fiona.klute@gmx.de>